backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
pandas
google-cloud-aiplatform
google-auth
google-genai
vertexai
